+++ sugar/test/src/time/weekdays_test.dart
@@ -3,23 +3,25 @@ import 'package:test/test.dart';
 
 void main() {
   const week = [1, 3, 5, 6];
-  const serialized = 53;
+  const bitfield = [true, false, true, false, true, true, false];
+  const encoded = 53;
 
-  test('pack', () => expect(Weekdays.encode(week), serialized));
+  test('encode', () => expect(Weekdays.encode(week), encoded));
 
-  test('pack more than 7 days', () => expect(
+  test('encode more than 7 days', () => expect(
       () => Weekdays.encode([1, 2, 3, 4, 5, 6, 7, 8]),
       throwsA(predicate<AssertionError>((e) => e.message == 'Number of days is "8", should be less than 7'))
   ));
-  test('pack invalid day', () => expect(
+
+  test('encode invalid day', () => expect(
       () => Weekdays.encode([9]),
       throwsA(predicate<AssertionError>((e) => e.message == 'Invalid day in week, should be between 1 and 7'))
   ));
 
 
-  test('unpack', () => expect(Weekdays.decode(serialized), week));
+  test('decode', () => expect(Weekdays.decode(encoded), week));
 
-  test('unpack invalid number', () {
+  test('decode invalid number', () {
     expect(
       () => Weekdays.decode(-1),
       throwsA(predicate<AssertionError>((e) => e.message == 'Packed days is "-1", should be between 0 and 127'))
@@ -31,5 +33,8 @@ void main() {
     );
   });
 
-  test('pack & unpack', () => expect(Weekdays.decode(Weekdays.encode(week)), week));
+  test('encode & decode', () => expect(Weekdays.decode(Weekdays.encode(week)), week));
+
+  test('parse', () => expect(Weekdays.parse(encoded), bitfield));
+
 }